add test and refactor Release into struct
